+++ packages/marketplace/src/components/pages/app-management/__tests__/apps-management.test.tsx
@@ -5,7 +5,7 @@ import { MemoryRouter } from 'react-router'
 import configureStore from 'redux-mock-store'
 import * as ReactRedux from 'react-redux'
 import routes from '@/constants/routes'
-import ClientAppsManagement, { handleOnChange, handleOnSettingClick } from '../client-apps-management'
+import ClientAppsManagement, { handleOnChange, handleOnSettingClick } from '../apps-management'
 import Routes from '@/constants/routes'
 import { ReduxState } from '@/types/core'
 import { getMockRouterProps } from '@/utils/mock-helper'
