+++ packages/marketplace/src/components/pages/app-management/index.ts
@@ -1,0 +1,1 @@
+export { default } from './apps-management'
