+++ packages/marketplace/src/components/pages/apps/index.ts
@@ -1,0 +1,1 @@
+export { default } from './apps'
