--- packages/marketplace/src/components/pages/client-app-management/index.ts
@@ -1,2 +1,0 @@
-import ClientAppsManagement from './client-apps-management'
-export default ClientAppsManagement
