--- packages/marketplace/src/components/pages/client/index.ts
@@ -1,2 +1,0 @@
-import Client from './client'
-export default Client
