+++ packages/marketplace/src/components/pages/settings/__tests__/setting.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { mount } from 'enzyme'
-import { ClientSettingsPage, handleLogout } from '../setting'
+import { Settings, handleLogout } from '../setting'
 import { authLogout } from '@/actions/auth'
 import { Provider } from 'react-redux'
 import { MemoryRouter } from 'react-router'
@@ -8,7 +8,7 @@ import Routes from '@/constants/routes'
 import configureStore from 'redux-mock-store'
 import appState from '@/reducers/__stubs__/app-state'
 
-describe('ClientSettingsPage', () => {
+describe('Settings', () => {
   it('should match snapshot', () => {
     const mockStore = configureStore()
     const store = mockStore(appState)
@@ -17,7 +17,7 @@ describe('ClientSettingsPage', () => {
       mount(
         <Provider store={store}>
           <MemoryRouter initialEntries={[{ pathname: Routes.SETTINGS, key: 'clientSettingsRoute' }]}>
-            <ClientSettingsPage />
+            <Settings />
           </MemoryRouter>
         </Provider>,
       ),
