+++ packages/marketplace/src/core/router.tsx
@@ -11,12 +11,12 @@ import { PortalProvider } from '@reapit/elements'
 export const history = createBrowserHistory()
 const Authentication = React.lazy(() => catchChunkError(() => import('../components/pages/authentication')))
 const Login = React.lazy(() => catchChunkError(() => import('../components/pages/login')))
-const Client = React.lazy(() => catchChunkError(() => import('../components/pages/client')))
+const Apps = React.lazy(() => catchChunkError(() => import('../components/pages/apps')))
 const AppDetail = React.lazy(() => catchChunkError(() => import('../components/pages/app-detail/client')))
 const WelcomePage = React.lazy(() => catchChunkError(() => import('../components/pages/welcome')))
 const InstalledApps = React.lazy(() => catchChunkError(() => import('../components/pages/installed-apps')))
 const Setting = React.lazy(() => catchChunkError(() => import('../components/pages/settings')))
-const AppsManagement = React.lazy(() => catchChunkError(() => import('../components/pages/client-app-management')))
+const AppsManagement = React.lazy(() => catchChunkError(() => import('../components/pages/app-management')))
 const Help = React.lazy(() => catchChunkError(() => import('../components/pages/help')))
 
 const Router = () => {
@@ -33,7 +33,7 @@ const Router = () => {
                 <PrivateRoute allow="CLIENT" path={Routes.AUTHENTICATION_LOGIN_TYPE} component={Authentication} />
                 <PrivateRoute allow="CLIENT" path={Routes.INSTALLED_APPS} component={InstalledApps} fetcher exact />
                 <PrivateRoute allow="CLIENT" path={Routes.MY_APPS} component={AppsManagement} fetcher exact />
-                <PrivateRoute allow="CLIENT" path={Routes.APPS} component={Client} exact fetcher />
+                <PrivateRoute allow="CLIENT" path={Routes.APPS} component={Apps} exact fetcher />
                 <PrivateRoute allow="CLIENT" path={Routes.APP_DETAIL} component={AppDetail} exact fetcher />
                 <PrivateRoute allow="CLIENT" path={Routes.APP_DETAIL_MANAGE} component={AppDetail} exact fetcher />
                 <PrivateRoute allow="CLIENT" path={Routes.WELCOME} component={WelcomePage} exact />
